Refactored how the templates are handled in the K8s backend.
